Fix nvidia driver rebinding nvidia-modprobe doesn't seem to load the driver anymore if there are no available GPUs but as soon as an available GPU exists it will eat it! additionally, loading driver on boot prevents it from being rebound, even though nvidia-smi shows nothing is using it so resort to binding as-needed

+++ vm/win10-pci.sh
@@ -36,7 +36,13 @@ rebind() {
     if [ "$driver" = "vfio-pci" ]; then
         echo $vendor $device > /sys/bus/pci/drivers/$driver/new_id
     else
-        echo $dev > /sys/bus/pci/drivers/$driver/bind
+        if [ "$driver" = "nvidia" ] && [ ! -e /sys/bus/pci/drivers/nvidia ]; then
+            # Special nvidia handling, version 495 nvidia-modprobe does not seem to create system files unless an
+            # unbound GPU is available but it does automatically eat the unbound GPU!
+            nvidia-modprobe && sleep 5
+        else
+            echo $dev > /sys/bus/pci/drivers/$driver/bind
+        fi
     fi
 }
